+++ src/components/content/loading-spinner/loading-spinner.cy.js
@@ -1,0 +1,12 @@
+import LoadingSpinner from "./loading-spinner.vue";
+import { createMount } from "@cypress/support/mount";
+
+const mount = createMount(LoadingSpinner);
+
+describe("loading-spinner", () => {
+	it("A component is rendered", () => {
+		mount();
+
+		cy.getByData("loading-spinner").shouldBeVisible();
+	});
+});