+++ src/components/content/loading-spinner/loading-spinner.vue
@@ -1,0 +1,3 @@
+<template>
+	<icon-loading class="spin text-grey-500 dark:text-white/60" data-test="loading-spinner" />
+</template>